Hand over the particle buffer and publish decals from Sim::Sync
Two more lanes of FUN_007474B0.

0x00747693..0x007476DB, the particle buffer. It is handed over rather than
shared: the packet takes the pointer and the sim is left with none, so the next
beat starts a fresh buffer. The decompile spells that out as a raw px copy, an
add-ref on the incoming control block, a weak_release of the outgoing one, and
then `[ebx+994h] = 0` / `[ebx+998h] = 0` with a full release -- an assignment
followed by a reset. The offsets confirm the model we already had:
Sim::mParticleBuffer at Sim+0x0994 (two reserved dwords after mCommandDB at
0x0988) and SSyncData::mParticleBuffer at +0x1C8, from `[esi+1CCh]` being the
control-block half.

0x007476EF, the decals. CDecalBuffer::SwapVectors (0x00779BB0) publishes both
lanes in one call -- visible decals into mAddDecals, pending hide ids into
mRemoveDecals. That function was fully recovered and had no caller anywhere in
the tree; this is its caller. CWldSession::DoBeat already hands both lanes to
CDecalManager::AddDecals / RemoveDecals, so it has been draining empty vectors.

Wiring it required settling one signedness disagreement, because a swap needs
both sides to name one type. The chain is SDecalInfo::mObj (std::uint32_t, +0x84)
-> AllocateDecalObjectId (std::uint32_t) -> CDecalBuffer::mPendingHideObjectIds
(std::uint32_t) -> the swap -> SSyncData::mRemoveDecals -> CDecalManager::
RemoveDecals. The value is unsigned at its source and at every producer, so the
two consumer-side declarations were the mistyped ones: SSyncData::mRemoveDecals
and CDecalManager::RemoveDecals both move to std::uint32_t. CWldTerrainDecal::
mRuntimeHandle stays std::int32_t -- it has its own offset assert and other
users -- so the one comparison against it is now explicit.

No offsets move: all of these are the same width either way.

tucheck clean on Sim.cpp, CWldSplat.cpp, CDecalBuffer.cpp and CWldSession.cpp.

+++ src/sdk/moho/terrain/splat/CWldSplat.cpp
@@ -483,11 +483,11 @@ namespace moho
    * Scans all active decals for each requested runtime handle and marks
    * matching decals for deferred removal.
    */
-  void CDecalManager::RemoveDecals(const msvc8::vector<std::int32_t>& decalHandles)
+  void CDecalManager::RemoveDecals(const msvc8::vector<std::uint32_t>& decalHandles)
   {
-    for (const std::int32_t handle : decalHandles) {
+    for (const std::uint32_t handle : decalHandles) {
       for (CWldTerrainDecal* const decal : mDecals) {
-        if (decal->mRuntimeHandle == handle) {
+        if (static_cast<std::uint32_t>(decal->mRuntimeHandle) == handle) {
           decal->mRemoveTick = 1;
           break;
         }
